pyproject.toml
[project]
name = "employee-catalog-django-ddd"
version = "0.1.0"
description = ""
authors = [
    {name = "chapppington",email = "stas.mega.96@bk.ru"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "django (>=5.2.8,<6.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "django-ninja (>=1.4.5,<2.0.0)",
    "punq (>=0.7.0,<0.8.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.4.0"
pytest = "^9.0.0"
isort = "^7.0.0"
faker = "^37.12.0"
pytest-django = "^4.11.1"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.core.project.settings.local.local"


[tool.isort]
multi_line_output=3
include_trailing_comma=true
line_length=120
lines_after_imports=2
lines_between_sections=1
lines_between_types=0
force_alphabetical_sort_within_sections=true
ensure_newline_before_comments=true
combine_as_imports=true
combine_star=true
skip_gitignore=true
balanced_wrapping=true
force_grid_wrap=2
src_paths=["core", ]
skip_glob=[
    ".git",
    ".idea",
    ".vscode",
    "venv",
    ".venv",
]
known_django=["django","ninja"]
known_first_party=["core"]
sections=[
    "FUTURE",
    "STDLIB",
    "DJANGO",
    "THIRDPARTY",
    "FIRSTPARTY",
    "LOCALFOLDER",
]